refactor: rename package name screening -> reservation
