+++ app/src/main/java/woowacourse/movie/contract/reservation/ScreeningContract.kt
@@ -1,0 +1,17 @@
+package woowacourse.movie.contract.reservation
+
+import woowacourse.movie.domain.reservation.Screening
+
+interface ScreeningContract {
+    interface Presenter {
+        fun presentScreenings()
+
+        fun selectScreening(screening: Screening)
+    }
+
+    interface View {
+        fun setScreenings(screenings: List<Screening>)
+
+        fun navigateToReservationScreen(screening: Screening)
+    }
+}
