--- app/src/main/java/woowacourse/movie/contract/screening/ReservationContract.kt
@@ -1,55 +1,0 @@
-package woowacourse.movie.contract.screening
-
-import java.time.LocalDate
-import java.time.LocalTime
-
-interface ReservationContract {
-    interface Presenter {
-        fun presentPoster()
-
-        fun presentTitle()
-
-        fun presentPeriod()
-
-        fun presentRunningTime()
-
-        fun presentDates()
-
-        fun presentTimes(date: LocalDate)
-
-        fun plusTicketCount(currentCount: Int)
-
-        fun minusTicketCount(currentCount: Int)
-
-        fun tryReservation()
-
-        fun confirmReservation()
-    }
-
-    interface View {
-        fun setPoster(movieId: Int)
-
-        fun setTitle(title: String)
-
-        fun setPeriod(
-            startYear: Int,
-            startMonth: Int,
-            startDay: Int,
-            endYear: Int,
-            endMonth: Int,
-            endDay: Int,
-        )
-
-        fun setRunningTime(runningTime: Int)
-
-        fun setDates(dates: List<LocalDate>)
-
-        fun setTimes(times: List<LocalTime>)
-
-        fun setTicketCount(count: Int)
-
-        fun showConfirmDialog()
-
-        fun navigateToTicketScreen(title: String)
-    }
-}
