--- app/src/main/java/woowacourse/movie/contract/screening/ScreeningContract.kt
@@ -1,17 +1,0 @@
-package woowacourse.movie.contract.screening
-
-import woowacourse.movie.domain.screening.Screening
-
-interface ScreeningContract {
-    interface Presenter {
-        fun presentScreenings()
-
-        fun selectScreening(screening: Screening)
-    }
-
-    interface View {
-        fun setScreenings(screenings: List<Screening>)
-
-        fun navigateToReservationScreen(screening: Screening)
-    }
-}
