--- app/src/main/java/woowacourse/movie/data/screening/Movies.kt
@@ -1,33 +1,0 @@
-package woowacourse.movie.data.screening
-
-import woowacourse.movie.domain.screening.Movie
-
-class Movies {
-    val harryPotterPhilosopersStone =
-        Movie(
-            0,
-            "해리 포터와 마법사의 돌",
-            152,
-        )
-
-    val harryPotterChamberOfSecrets =
-        Movie(
-            1,
-            "해리 포터와 비밀의 방",
-            162,
-        )
-
-    val harryPotterPrisonerOfAzkaban =
-        Movie(
-            2,
-            "해리 포터와 아즈카반의 죄수",
-            141,
-        )
-
-    val harryPotterGobletOfFire =
-        Movie(
-            3,
-            "해리 포터와 불의 잔",
-            157,
-        )
-}
