+++ app/src/main/java/woowacourse/movie/domain/reservation/Advertisement.kt
@@ -1,0 +1,6 @@
+package woowacourse.movie.domain.reservation
+
+data class Advertisement(
+    val id: Int,
+    val description: String? = null,
+) : ScreeningContent
