+++ app/src/main/java/woowacourse/movie/domain/reservation/Advertisements.kt
@@ -1,0 +1,14 @@
+package woowacourse.movie.domain.reservation
+
+class Advertisements(
+    private val value: List<Advertisement>,
+) {
+    private val size = value.size
+    private var position = 0
+
+    fun fetchAdvertisement(): Advertisement {
+        val advertisement: Advertisement = value[position]
+        position = position.plus(1) % size
+        return advertisement
+    }
+}
