+++ app/src/main/java/woowacourse/movie/domain/reservation/Movie.kt
@@ -1,0 +1,9 @@
+package woowacourse.movie.domain.reservation
+
+import java.io.Serializable
+
+data class Movie(
+    val id: Int,
+    val title: String,
+    val runningTime: Int,
+) : Serializable
