+++ app/src/main/java/woowacourse/movie/domain/reservation/ScreeningContent.kt
@@ -1,0 +1,3 @@
+package woowacourse.movie.domain.reservation
+
+interface ScreeningContent
