--- app/src/main/java/woowacourse/movie/domain/screening/Advertisement.kt
@@ -1,6 +1,0 @@
-package woowacourse.movie.domain.screening
-
-data class Advertisement(
-    val id: Int,
-    val description: String? = null,
-) : ScreeningContent
