--- app/src/main/java/woowacourse/movie/domain/screening/Movie.kt
@@ -1,9 +1,0 @@
-package woowacourse.movie.domain.screening
-
-import java.io.Serializable
-
-data class Movie(
-    val id: Int,
-    val title: String,
-    val runningTime: Int,
-) : Serializable
