--- app/src/main/java/woowacourse/movie/domain/screening/ScreeningContent.kt
@@ -1,3 +1,0 @@
-package woowacourse.movie.domain.screening
-
-interface ScreeningContent
