--- app/src/main/java/woowacourse/movie/view/screening/ShowReservationConfirmDialog.kt
@@ -1,33 +1,0 @@
-package woowacourse.movie.view.screening
-
-import android.content.Context
-import android.content.DialogInterface
-import androidx.appcompat.app.AlertDialog
-
-class ShowReservationConfirmDialog(
-    context: Context,
-) {
-    private val builder = AlertDialog.Builder(context)
-
-    operator fun invoke(
-        title: String,
-        message: String,
-        positiveButtonText: String,
-        positiveButtonAction: (dialog: DialogInterface, which: Int) -> Unit,
-        negativeButtonText: String,
-        negativeButtonAction: (dialog: DialogInterface, which: Int) -> Unit,
-        cancelable: Boolean = false,
-    ) {
-        builder
-            .setTitle(title)
-            .setMessage(message)
-            .setPositiveButton(
-                positiveButtonText,
-                positiveButtonAction,
-            ).setNegativeButton(
-                negativeButtonText,
-                negativeButtonAction,
-            ).setCancelable(cancelable)
-            .show()
-    }
-}
