--- app/src/test/java/woowacourse/movie/presenter/screening/ScreeningPresenterTest.kt
@@ -1,127 +1,0 @@
-package woowacourse.movie.presenter.screening
-
-import io.mockk.Runs
-import io.mockk.every
-import io.mockk.just
-import io.mockk.mockk
-import io.mockk.verify
-import org.junit.jupiter.api.BeforeEach
-import org.junit.jupiter.api.Test
-import woowacourse.movie.contract.screening.ScreeningContract
-import woowacourse.movie.data.screening.FakeScreenings
-import woowacourse.movie.domain.screening.Movie
-import woowacourse.movie.domain.screening.Screening
-import java.time.LocalDate
-
-class ScreeningPresenterTest {
-    private lateinit var presenter: ScreeningContract.Presenter
-    private lateinit var view: ScreeningContract.View
-
-    @BeforeEach
-    fun setUp() {
-        view = mockk()
-        presenter =
-            ScreeningPresenter(
-                view,
-                FakeScreenings(
-                    listOf(
-                        Screening(
-                            Movie(
-                                0,
-                                "해리 포터와 마법사의 돌",
-                                152,
-                            ),
-                            LocalDate.of(2025, 4, 1),
-                            LocalDate.of(2025, 4, 25),
-                        ),
-                    ),
-                ),
-            )
-    }
-
-    @Test
-    fun `영화 목록을 보여줄 수 있다`() {
-        // given
-        every {
-            view.setScreenings(
-                listOf(
-                    Screening(
-                        Movie(
-                            0,
-                            "해리 포터와 마법사의 돌",
-                            152,
-                        ),
-                        LocalDate.of(2025, 4, 1),
-                        LocalDate.of(2025, 4, 25),
-                    ),
-                ),
-            )
-        } just Runs
-
-        // when
-        presenter.presentScreenings()
-
-        // then
-        verify {
-            view.setScreenings(
-                listOf(
-                    Screening(
-                        Movie(
-                            0,
-                            "해리 포터와 마법사의 돌",
-                            152,
-                        ),
-                        LocalDate.of(2025, 4, 1),
-                        LocalDate.of(2025, 4, 25),
-                    ),
-                ),
-            )
-        }
-    }
-
-    @Test
-    fun `예매할 영화를 선택할 수 있다`() {
-        // given
-        every {
-            view.navigateToReservationScreen(
-                Screening(
-                    Movie(
-                        0,
-                        "해리 포터와 마법사의 돌",
-                        152,
-                    ),
-                    LocalDate.of(2025, 4, 1),
-                    LocalDate.of(2025, 4, 25),
-                ),
-            )
-        } just Runs
-
-        // when
-        presenter.selectScreening(
-            Screening(
-                Movie(
-                    0,
-                    "해리 포터와 마법사의 돌",
-                    152,
-                ),
-                LocalDate.of(2025, 4, 1),
-                LocalDate.of(2025, 4, 25),
-            ),
-        )
-
-        // then
-        verify {
-            view.navigateToReservationScreen(
-                Screening(
-                    Movie(
-                        0,
-                        "해리 포터와 마법사의 돌",
-                        152,
-                    ),
-                    LocalDate.of(2025, 4, 1),
-                    LocalDate.of(2025, 4, 25),
-                ),
-            )
-        }
-    }
-}
